Beginning transition to p4python
The original Perforce.tmbundle was a very simple wrapper for "p4
$COMMAND $TM_FILEPATH"  It only worked on the file you were currently
editing (you couldn't select others in the project sidebar).  This
version uses p4python instead of the command line p4 client.  It
iterates over all the selected files, ignores files not in your Perforce
workspace, and prepares selected directories for Perforce by adding
"/..." to the ends of their paths before forwarding a command to the
Perforce server.