Converting Open in new Document to Show as HTML
TextMate runs Commands with `Create New Document` synchronously, but
`Show as HTML` asynchronously.  (TextMate will hang until the script
returns, so if you call "mate -w" in the script, you've essentially just
entered an infinite loop.)  I've converted all instances of `Create New
Document` to `Show as HTML` to allow changelists to be authored in
TextMate instead of in the clunky CommitWindow.